refactor: rename error paths
